QtNetwork: module updated
Modified:
source/QtNetwork/QAbstractNetworkCacheClass.prg
source/QtNetwork/QAbstractSocketClass.prg
source/QtNetwork/QFtpClass.prg
source/QtNetwork/QHostInfoClass.prg
source/QtNetwork/QHttpMultiPartClass.prg
source/QtNetwork/QHttpPartClass.prg
source/QtNetwork/QLocalServerClass.prg
source/QtNetwork/QLocalSocketClass.prg
source/QtNetwork/QNetworkAccessManagerClass.prg
source/QtNetwork/QNetworkAddressEntryClass.prg
source/QtNetwork/QNetworkCacheMetaDataClass.prg
source/QtNetwork/QNetworkConfigurationClass.prg
source/QtNetwork/QNetworkConfigurationManagerClass.prg
source/QtNetwork/QNetworkCookieClass.prg
source/QtNetwork/QNetworkCookieJarClass.prg
source/QtNetwork/QNetworkDiskCacheClass.prg
source/QtNetwork/QNetworkInterfaceClass.prg
source/QtNetwork/QNetworkProxyClass.prg
source/QtNetwork/QNetworkProxyFactoryClass.prg
source/QtNetwork/QNetworkProxyQueryClass.prg
source/QtNetwork/QNetworkReplyClass.prg
source/QtNetwork/QNetworkRequestClass.prg
source/QtNetwork/QNetworkSessionClass.prg
source/QtNetwork/QSslCertificateClass.prg
source/QtNetwork/QSslCipherClass.prg
source/QtNetwork/QSslConfigurationClass.prg
source/QtNetwork/QSslErrorClass.prg
source/QtNetwork/QSslKeyClass.prg
source/QtNetwork/QSslSocketClass.prg
source/QtNetwork/QTcpServerClass.prg
source/QtNetwork/QTcpSocketClass.prg
source/QtNetwork/QUrlInfoClass.prg

+++ source/QtNetwork/QFtpClass.prg
@@ -15,8 +15,6 @@ REQUEST QBYTEARRAY
 
 CLASS QFtp INHERIT QObject
 
-   DATA self_destruction INIT .F.
-
    METHOD new
    METHOD delete
    METHOD bytesAvailable
@@ -47,6 +45,7 @@ CLASS QFtp INHERIT QObject
    METHOD setTransferMode
    METHOD state
    METHOD abort
+
    METHOD onCommandFinished
    METHOD onCommandStarted
    METHOD onDataTransferProgress
@@ -55,6 +54,7 @@ CLASS QFtp INHERIT QObject
    METHOD onRawCommandReply
    METHOD onReadyRead
    METHOD onStateChanged
+
    DESTRUCTOR destroyObject
 
 END CLASS
@@ -78,8 +78,15 @@ QFtp ( QObject * parent = 0 )
 */
 HB_FUNC_STATIC( QFTP_NEW )
 {
-  QFtp * o = new QFtp ( OPQOBJECT(1,0) );
-  _qt4xhb_storePointerAndFlag ( o, false );
+  if( ISBETWEEN(0,1) && (ISQOBJECT(1)||ISNIL(1)) )
+  {
+    QFtp * o = new QFtp ( OPQOBJECT(1,0) );
+    _qt4xhb_storePointerAndFlag( o, false );
+  }
+  else
+  {
+    hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+  }
 }
 
 HB_FUNC_STATIC( QFTP_DELETE )
@@ -105,9 +112,17 @@ qint64 bytesAvailable () const
 HB_FUNC_STATIC( QFTP_BYTESAVAILABLE )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    RQINT64( obj->bytesAvailable () );
+    if( ISNUMPAR(0) )
+    {
+      RQINT64( obj->bytesAvailable () );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -117,9 +132,17 @@ int cd ( const QString & dir )
 HB_FUNC_STATIC( QFTP_CD )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    RINT( obj->cd ( PQSTRING(1) ) );
+    if( ISNUMPAR(1) && ISCHAR(1) )
+    {
+      RINT( obj->cd ( PQSTRING(1) ) );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -129,10 +152,19 @@ void clearPendingCommands ()
 HB_FUNC_STATIC( QFTP_CLEARPENDINGCOMMANDS )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    obj->clearPendingCommands ();
+    if( ISNUMPAR(0) )
+    {
+      obj->clearPendingCommands ();
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
+
   hb_itemReturn( hb_stackSelfItem() );
 }
 
@@ -142,9 +174,17 @@ int close ()
 HB_FUNC_STATIC( QFTP_CLOSE )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    RINT( obj->close () );
+    if( ISNUMPAR(0) )
+    {
+      RINT( obj->close () );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -154,9 +194,17 @@ int connectToHost ( const QString & host, quint16 port = 21 )
 HB_FUNC_STATIC( QFTP_CONNECTTOHOST )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    RINT( obj->connectToHost ( PQSTRING(1), OPQUINT16(2,21) ) );
+    if( ISBETWEEN(1,2) && ISCHAR(1) && ISOPTNUM(2) )
+    {
+      RINT( obj->connectToHost ( PQSTRING(1), OPQUINT16(2,21) ) );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -166,9 +214,17 @@ Command currentCommand () const
 HB_FUNC_STATIC( QFTP_CURRENTCOMMAND )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    hb_retni( (int) obj->currentCommand () );
+    if( ISNUMPAR(0) )
+    {
+      RENUM( obj->currentCommand () );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -178,10 +234,18 @@ QIODevice * currentDevice () const
 HB_FUNC_STATIC( QFTP_CURRENTDEVICE )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    QIODevice * ptr = obj->currentDevice ();
-    _qt4xhb_createReturnClass ( ptr, "QIODEVICE" );
+    if( ISNUMPAR(0) )
+    {
+      QIODevice * ptr = obj->currentDevice ();
+      _qt4xhb_createReturnQObjectClass ( ptr, "QIODEVICE" );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -191,9 +255,17 @@ int currentId () const
 HB_FUNC_STATIC( QFTP_CURRENTID )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    RINT( obj->currentId () );
+    if( ISNUMPAR(0) )
+    {
+      RINT( obj->currentId () );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -203,9 +275,17 @@ Error error () const
 HB_FUNC_STATIC( QFTP_ERROR )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    hb_retni( (int) obj->error () );
+    if( ISNUMPAR(0) )
+    {
+      RENUM( obj->error () );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -235,11 +315,17 @@ int get ( const QString & file, QIODevice * dev = 0, TransferType type = Binary 
 HB_FUNC_STATIC( QFTP_GET )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    QIODevice * par2 = ISNIL(2)? 0 : (QIODevice *) hb_itemGetPtr( hb_objSendMsg( hb_param(2, HB_IT_OBJECT ), "POINTER", 0 ) );
-    int par3 = ISNIL(3)? (int) QFtp::Binary : hb_parni(3);
-    RINT( obj->get ( PQSTRING(1), par2, (QFtp::TransferType) par3 ) );
+    if( ISBETWEEN(1,3) && ISCHAR(1) && (ISQIODEVICE(2)||ISNIL(2)) && ISOPTNUM(3) )
+    {
+      RINT( obj->get ( PQSTRING(1), OPQIODEVICE(2,0), ISNIL(3)? (QFtp::TransferType) QFtp::Binary : (QFtp::TransferType) hb_parni(3) ) );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -249,9 +335,17 @@ bool hasPendingCommands () const
 HB_FUNC_STATIC( QFTP_HASPENDINGCOMMANDS )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    RBOOL( obj->hasPendingCommands () );
+    if( ISNUMPAR(0) )
+    {
+      RBOOL( obj->hasPendingCommands () );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -261,9 +355,17 @@ int list ( const QString & dir = QString() )
 HB_FUNC_STATIC( QFTP_LIST )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    RINT( obj->list ( OPQSTRING(1,QString()) ) );
+    if( ISNUMPAR(1) && ISOPTCHAR(1) )
+    {
+      RINT( obj->list ( OPQSTRING(1,QString()) ) );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -273,9 +375,17 @@ int login ( const QString & user = QString(), const QString & password = QString
 HB_FUNC_STATIC( QFTP_LOGIN )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    RINT( obj->login ( OPQSTRING(1,QString()), OPQSTRING(2,QString()) ) );
+    if( ISNUMPAR(2) && ISOPTCHAR(1) && ISOPTCHAR(2) )
+    {
+      RINT( obj->login ( OPQSTRING(1,QString()), OPQSTRING(2,QString()) ) );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -285,9 +395,17 @@ int mkdir ( const QString & dir )
 HB_FUNC_STATIC( QFTP_MKDIR )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    RINT( obj->mkdir ( PQSTRING(1) ) );
+    if( ISNUMPAR(1) && ISCHAR(1) )
+    {
+      RINT( obj->mkdir ( PQSTRING(1) ) );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -297,10 +415,17 @@ int put ( QIODevice * dev, const QString & file, TransferType type = Binary )
 HB_FUNC_STATIC( QFTP_PUT1 )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    int par3 = ISNIL(3)? (int) QFtp::Binary : hb_parni(3);
-    RINT( obj->put ( PQIODEVICE(1), PQSTRING(2), (QFtp::TransferType) par3 ) );
+    if( ISBETWEEN(2,3) && ISQIODEVICE(1) && ISCHAR(2) && ISOPTNUM(3) )
+    {
+      RINT( obj->put ( PQIODEVICE(1), PQSTRING(2), ISNIL(3)? (QFtp::TransferType) QFtp::Binary : (QFtp::TransferType) hb_parni(3) ) );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -310,10 +435,17 @@ int put ( const QByteArray & data, const QString & file, TransferType type = Bin
 HB_FUNC_STATIC( QFTP_PUT2 )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    int par3 = ISNIL(3)? (int) QFtp::Binary : hb_parni(3);
-    RINT( obj->put ( *PQBYTEARRAY(1), PQSTRING(2), (QFtp::TransferType) par3 ) );
+    if( ISBETWEEN(2,3) && ISQBYTEARRAY(1) && ISCHAR(2) && ISOPTNUM(3) )
+    {
+      RINT( obj->put ( *PQBYTEARRAY(1), PQSTRING(2), ISNIL(3)? (QFtp::TransferType) QFtp::Binary : (QFtp::TransferType) hb_parni(3) ) );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -338,9 +470,17 @@ int rawCommand ( const QString & command )
 HB_FUNC_STATIC( QFTP_RAWCOMMAND )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    RINT( obj->rawCommand ( PQSTRING(1) ) );
+    if( ISNUMPAR(1) && ISCHAR(1) )
+    {
+      RINT( obj->rawCommand ( PQSTRING(1) ) );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -350,10 +490,17 @@ qint64 read ( char * data, qint64 maxlen )
 HB_FUNC_STATIC( QFTP_READ )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    char * par1 = (char *) hb_itemGetPtr( hb_objSendMsg( hb_param(1, HB_IT_OBJECT ), "POINTER", 0 ) );
-    RQINT64( obj->read ( par1, PQINT64(2) ) );
+    if( ISNUMPAR(2) && ISCHAR(1) && ISNUM(2) )
+    {
+      RQINT64( obj->read ( (char *) hb_parc(1), PQINT64(2) ) );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -363,10 +510,18 @@ QByteArray readAll ()
 HB_FUNC_STATIC( QFTP_READALL )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    QByteArray * ptr = new QByteArray( obj->readAll () );
-    _qt4xhb_createReturnClass ( ptr, "QBYTEARRAY" );
+    if( ISNUMPAR(0) )
+    {
+      QByteArray * ptr = new QByteArray( obj->readAll () );
+      _qt4xhb_createReturnClass ( ptr, "QBYTEARRAY", true );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -376,9 +531,17 @@ int remove ( const QString & file )
 HB_FUNC_STATIC( QFTP_REMOVE )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    RINT( obj->remove ( PQSTRING(1) ) );
+    if( ISNUMPAR(1) && ISCHAR(1) )
+    {
+      RINT( obj->remove ( PQSTRING(1) ) );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -388,9 +551,17 @@ int rename ( const QString & oldname, const QString & newname )
 HB_FUNC_STATIC( QFTP_RENAME )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    RINT( obj->rename ( PQSTRING(1), PQSTRING(2) ) );
+    if( ISNUMPAR(2) && ISCHAR(1) && ISCHAR(2) )
+    {
+      RINT( obj->rename ( PQSTRING(1), PQSTRING(2) ) );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -400,9 +571,17 @@ int rmdir ( const QString & dir )
 HB_FUNC_STATIC( QFTP_RMDIR )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    RINT( obj->rmdir ( PQSTRING(1) ) );
+    if( ISNUMPAR(1) && ISCHAR(1) )
+    {
+      RINT( obj->rmdir ( PQSTRING(1) ) );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -412,9 +591,17 @@ int setProxy ( const QString & host, quint16 port )
 HB_FUNC_STATIC( QFTP_SETPROXY )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    RINT( obj->setProxy ( PQSTRING(1), PQUINT16(2) ) );
+    if( ISNUMPAR(2) && ISCHAR(1) && ISNUM(2) )
+    {
+      RINT( obj->setProxy ( PQSTRING(1), PQUINT16(2) ) );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -424,9 +611,17 @@ int setTransferMode ( TransferMode mode )
 HB_FUNC_STATIC( QFTP_SETTRANSFERMODE )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    RINT( obj->setTransferMode ( (QFtp::TransferMode) hb_parni(1) ) );
+    if( ISNUMPAR(1) && ISNUM(1) )
+    {
+      RINT( obj->setTransferMode ( (QFtp::TransferMode) hb_parni(1) ) );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -436,9 +631,17 @@ State state () const
 HB_FUNC_STATIC( QFTP_STATE )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    hb_retni( (int) obj->state () );
+    if( ISNUMPAR(0) )
+    {
+      RENUM( obj->state () );
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
 }
 
@@ -448,10 +651,19 @@ void abort ()
 HB_FUNC_STATIC( QFTP_ABORT )
 {
   QFtp * obj = (QFtp *) hb_itemGetPtr( hb_objSendMsg( hb_stackSelfItem(), "POINTER", 0 ) );
+
   if( obj )
   {
-    obj->abort ();
+    if( ISNUMPAR(0) )
+    {
+      obj->abort ();
+    }
+    else
+    {
+      hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
+    }
   }
+
   hb_itemReturn( hb_stackSelfItem() );
 }
